+++ src/components/Button/Button_snapshot.spec.js
@@ -37,6 +37,32 @@ describe('Button Component', () => {
       );
       expect(wrapper5).toMatchSnapshot();
 
+      const wrapper6 = shallow(
+        <Button appearance="link">Push Me</Button>
+      );
+      expect(wrapper6).toMatchSnapshot();
+
+      const wrapper7 = shallow(
+        <Button appearance="link" size="small">
+          Push Me
+        </Button>
+      );
+      expect(wrapper7).toMatchSnapshot();
+
+      const wrapper8 = shallow(
+        <Button appearance="link" disabled>
+          Don't Push Me
+        </Button>
+      );
+      expect(wrapper8).toMatchSnapshot();
+
+      const wrapper9 = shallow(
+        <Button appearance="primary" renderAs="span">
+          I'm not a button
+        </Button>
+      );
+      expect(wrapper9).toMatchSnapshot();
+
     });
   });
 });