+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+### Changed
+
+- Button now has a `link` appearance.
+- Button now has an option to be rendered as any given HTML tag.
+
 ## [3.8.0] - 2019-02-26
 
 ### Added